renamed 'sensorlog-opt.js' to 'log-opt.js'
