+++ mysql/log-opt.js
